requirements.txt
absl-py
alabaster==0.7.16
appdirs==1.4.4
argon2-cffi
astunparse==1.6.3
attrs==21.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget
bleach==4.1.0
blinker
Bottleneck
Brotli==1.0.9
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cloudpickle==3.0.0
colorama==0.4.4
cryptography==3.4.8
cycler==0.11.0
Cython
decorator==4.4.2
defer
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.20.1
entrypoints==0.4
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.6.0
flake8==4.0.1
flatbuffers
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast
gitdb==4.0.11
GitPython==3.1.42
google-auth
google-auth-oauthlib
google-pasta
grpcio
html5lib==1.1
httplib2==0.20.2
idna
imageio
imagesize==1.4.1
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jax==0.4.14
jdcal
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.13.1
keras-tuner==1.4.7
keyring==23.5.0
kiwisolver==1.3.2
kt-legacy==1.0.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.38.0
lockfile==0.12.2
loguru==0.7.2
louis
lxml==4.8.0
lz4
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.4
nose==1.3.7
notebook==6.4.8
numba
numexpr==2.8.1
numpy
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
odfpy
olefile==0.46
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parameterized==0.8.1
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.3.0
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.17.2
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.1
pythran
pytools==2021.2.8
pytz==2022.1
pyxdg==0.27
PyYAML
pyzmq==22.3.0
reportlab==3.6.8
requests
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn
scipy==1.8.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sentry-sdk==1.41.0
setproctitle==1.3.3
simplejson==3.17.6
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
ssh-import-id==5.11
sympy==1.9
tables==3.7.0
tensorboard==2.13.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-probability==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
Theano
threadpoolctl==3.1.0
toml
torchvision==0.15.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
ufoLib2==0.13.1
unicodedata2==14.0.0
urllib3==2.2.1
wadllib==1.3.6
wandb==0.16.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==2.0.0
wrapt==1.13.3
xdg==5
xlwt==1.3.0
zipp==1.0.0